Build: update core module dependencies after refactoring
- Update core/data dependencies and imports
- Fix core/network circular dependency
- Update core/ui to depend on model module
- Update all imports to new package structure
