+++ core/data/src/commonMain/kotlin/pt/dourobats/app/core/data/di/DataModule.kt
@@ -5,9 +5,9 @@ import pt.dourobats.app.core.data.preferences.createDataStore
 import pt.dourobats.app.core.data.repository.AuthRepositoryImpl
 import pt.dourobats.app.core.data.repository.FakeTrainingRepository
 import pt.dourobats.app.core.data.repository.SettingsRepositoryImpl
-import pt.dourobats.app.core.domain.repository.AuthRepository
-import pt.dourobats.app.core.domain.repository.SettingsRepository
-import pt.dourobats.app.core.domain.repository.TrainingRepository
+import pt.dourobats.app.core.repository.AuthRepository
+import pt.dourobats.app.core.repository.SettingsRepository
+import pt.dourobats.app.core.repository.TrainingRepository
 
 val dataModule = module {
     single { createDataStore() }
