+++ core/ui/build.gradle.kts
@@ -7,7 +7,7 @@ plugins {
 kotlin {
     sourceSets {
         commonMain.dependencies {
-            implementation(projects.core.domain)
+            implementation(projects.core.model)
             api(libs.kotlinx.datetime)  // API because we expose kotlinx.datetime types in public API
         }
     }
