+++ core/ui/src/commonMain/kotlin/pt/dourobats/app/core/ui/localization/LanguageChanger.kt
@@ -1,7 +1,7 @@
 package pt.dourobats.app.core.ui.localization
 
 import androidx.compose.runtime.staticCompositionLocalOf
-import pt.dourobats.app.core.domain.model.Language
+import pt.dourobats.app.core.model.Language
 
 /**
  * CompositionLocal for the current language.
