+++ core/ui/src/iosMain/kotlin/pt/dourobats/app/core/ui/localization/LanguageChanger.ios.kt
@@ -1,7 +1,7 @@
 package pt.dourobats.app.core.ui.localization
 
 import platform.Foundation.NSUserDefaults
-import pt.dourobats.app.core.domain.model.Language
+import pt.dourobats.app.core.model.Language
 
 actual fun changeLanguage(language: Language) {
     // On iOS, update the AppleLanguages preference
